MAINT: Set page.rotate(angle: int) (#1092)
Closes #1021

+++ PyPDF2/_page.py
@@ -298,7 +298,7 @@ class PageObject(DictionaryObject):
         deprecate_with_replacement("createBlankPage", "create_blank_page")
         return PageObject.create_blank_page(pdf, width, height)
 
-    def rotate(self, angle: float) -> "PageObject":
+    def rotate(self, angle: int) -> "PageObject":
         """
         Rotate a page clockwise by increments of 90 degrees.
 
@@ -314,11 +314,11 @@ class PageObject(DictionaryObject):
         self[NameObject(PG.ROTATE)] = NumberObject(current_angle + angle)
         return self
 
-    def rotate_clockwise(self, angle: float) -> "PageObject":  # pragma: no cover
+    def rotate_clockwise(self, angle: int) -> "PageObject":  # pragma: no cover
         deprecate_with_replacement("rotate_clockwise", "rotate")
         return self.rotate(angle)
 
-    def rotateClockwise(self, angle: float) -> "PageObject":  # pragma: no cover
+    def rotateClockwise(self, angle: int) -> "PageObject":  # pragma: no cover
         """
         .. deprecated:: 1.28.0
 
@@ -327,7 +327,7 @@ class PageObject(DictionaryObject):
         deprecate_with_replacement("rotateClockwise", "rotate")
         return self.rotate(angle)
 
-    def rotateCounterClockwise(self, angle: float) -> "PageObject":  # pragma: no cover
+    def rotateCounterClockwise(self, angle: int) -> "PageObject":  # pragma: no cover
         """
         .. deprecated:: 1.28.0
